backend/requirements.txt
Django>=4.2,<6.0
django-cors-headers>=4.3
django-allauth>=0.57
google-auth>=2.20
requests>=2.31
python-dotenv>=1.0
PyJWT>=2.8
cryptography>=42.0
Pillow>=10.4
mysqlclient>=2.2
